+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## unreleased
 ### Added
 * Added support to xc3_viewer for loading multiple `.wimdo` or `.camdo` files.
+* Added `--bones` and `--bounds` debug shading args to xc3_viewer.
 
 ### Changed
 * Improved accuracy of render state and render pass assignments for Xenoblade X models.